Beacon.
Solves the question of what IP the phone company hands over to the drone.
It is assumed that we connect the drone via 3-/4-G connection and that it is
dynamic.

beacon.service needs to be enabled on the drone.
beacon.py needs to be linked to /usr/sbin/beacon.py